+++ adrestus-consensus/src/test/java/io/Adrestus/consensus/ConsensusCommitteeTimer4Test.java
@@ -76,6 +76,10 @@ public class ConsensusCommitteeTimer4Test {
 
     @BeforeAll
     public static void setup() throws Exception {
+        if (System.out.getClass().getName().contains("maven")) {
+            return;
+        }
+
         IDatabase<String, CommitteeBlock> db = new DatabaseFactory(String.class, CommitteeBlock.class).getDatabase(DatabaseType.ROCKS_DB, DatabaseInstance.COMMITTEE_BLOCK);
         db.delete_db();
         IDatabase<String, CommitteeBlock> database = new DatabaseFactory(String.class, CommitteeBlock.class).getDatabase(DatabaseType.ROCKS_DB, DatabaseInstance.COMMITTEE_BLOCK);
@@ -191,7 +195,7 @@ public class ConsensusCommitteeTimer4Test {
 
     @Test
     public void committe_test() throws Exception {
-        if (System.getenv("MAVEN_OPTS") != null) {
+        if (System.out.getClass().getName().contains("maven")) {
             return;
         }
         Socket socket = new Socket();
