+++ adrestus-consensus/src/test/java/io/Adrestus/consensus/ConsensusCommitteeTimerTest.java
@@ -57,6 +57,10 @@ public class ConsensusCommitteeTimerTest {
 
     @BeforeAll
     public static void setup() throws Exception {
+        if (System.out.getClass().getName().contains("maven")) {
+            return;
+        }
+
         IDatabase<String, CommitteeBlock> db = new DatabaseFactory(String.class, CommitteeBlock.class).getDatabase(DatabaseType.ROCKS_DB, DatabaseInstance.COMMITTEE_BLOCK);
         db.delete_db();
 
@@ -115,7 +119,7 @@ public class ConsensusCommitteeTimerTest {
 
     @Test
     public void committe_test() throws Exception {
-        if (System.getenv("MAVEN_OPTS") != null) {
+        if (System.out.getClass().getName().contains("maven")) {
             return;
         }
         Socket socket = new Socket();
