+++ adrestus-consensus/src/test/java/io/Adrestus/consensus/ConsensusCommitteeTimerTest3.java
@@ -65,6 +65,9 @@ public class ConsensusCommitteeTimerTest3 {
 
     @BeforeAll
     public static void setup() throws Exception {
+        if (System.out.getClass().getName().contains("maven")) {
+            return;
+        }
         IDatabase<String, CommitteeBlock> db = new DatabaseFactory(String.class, CommitteeBlock.class).getDatabase(DatabaseType.ROCKS_DB, DatabaseInstance.COMMITTEE_BLOCK);
         db.delete_db();
         IDatabase<String, CommitteeBlock> database = new DatabaseFactory(String.class, CommitteeBlock.class).getDatabase(DatabaseType.ROCKS_DB, DatabaseInstance.COMMITTEE_BLOCK);
@@ -162,7 +165,7 @@ public class ConsensusCommitteeTimerTest3 {
 
     @Test
     public void committe_test() throws Exception {
-        if (System.getenv("MAVEN_OPTS") != null) {
+        if (System.out.getClass().getName().contains("maven")) {
             return;
         }
         Socket socket = new Socket();
